+++ src/c_ext/objectives.h
@@ -1,9 +1,9 @@
 #ifndef OBJECTIVES_H
 #define OBJECTIVES_H
 
-double W(int A, int B, double *weights, int n);
+double W(long A, long B, double *weights, int n);
 
-double F(int A, int B, double *weights, int n);
+double F(long A, long B, double *weights, int n);
 
 int S(double *x, double **weights, int n);
